+++ map/mobile/android/src/ti/map/MapModule.java
@@ -29,6 +29,8 @@ public class MapModule extends KrollModule
 	public static final String PROPERTY_LEFT_PANE = "leftPane";
 	public static final String PROPERTY_RIGHT_PANE = "rightPane";
 	public static final String EVENT_PIN_CHANGE_DRAG_STATE = "pinchangedragstate";
+	public static final String PROPERTY_TILT_ENABLED = "tiltEnabled";
+	public static final String PROPERTY_ROTATE_ENABLED = "rotateEnabled";
 	
 	@Kroll.constant public static final int NORMAL_TYPE = GoogleMap.MAP_TYPE_NORMAL;
 	@Kroll.constant public static final int TERRAIN_TYPE = GoogleMap.MAP_TYPE_TERRAIN;